BulkLoader: Don't reuse PreparedStatement instances
When prepared statements are used with async execution, we must not
reuse PreparedStatement instances until the async operation is complete.
Therefore, drop the prepared statement ConcurrentHashMap altogether and
always build a new PreparedStatement instance. Please note that this
does *not* mean re-preparing the statement on server side, that's going
to be reused anyway automatically.